+++ contracts/mocks/MockPool.sol
@@ -8,19 +8,14 @@ import "./MockToken.sol";
 contract MockPool {
     address public token0;
     address public token1;
-    uint24 public constant fee = 3000; 
+    uint24 public constant fee = 3000;
 
     constructor(address _token0, address _token1) {
         token0 = _token0;
         token1 = _token1;
     }
 
-    function flash(
-        address recipient,
-        uint256 amount0,
-        uint256 amount1,
-        bytes calldata data
-    ) external {
+    function flash(address recipient, uint256 amount0, uint256 amount1, bytes calldata data) external {
         uint256 balance0Before = IERC20(token0).balanceOf(address(this));
         uint256 balance1Before = IERC20(token1).balanceOf(address(this));
 
@@ -49,23 +44,22 @@ contract MockPool {
     }
 
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint /*amountOutMin */,
+        uint256 amountIn,
+        uint256, /*amountOutMin */
         address[] calldata path,
         address to,
-        uint /*deadline*/
-    ) external returns (uint[] memory amounts) {
-        
+        uint256 /*deadline*/
+    ) external returns (uint256[] memory amounts) {
         address tokenIn = path[0];
         address tokenOut = path[path.length - 1];
 
         IERC20(tokenIn).transferFrom(msg.sender, address(this), amountIn);
 
-        uint256 amountOut = (amountIn * 1020) / 1000; 
+        uint256 amountOut = (amountIn * 1020) / 1000;
 
         MockToken(tokenOut).mint(to, amountOut);
 
-        amounts = new uint[](2);
+        amounts = new uint256[](2);
         amounts[0] = amountIn;
         amounts[1] = amountOut;
         return amounts;
